feat:任务规划增加全家指令按钮和手动介入按钮
