+++ app/renderer/src/main/src/pages/ai-agent/template/template.tsx
@@ -12,10 +12,7 @@ import React, {
 import {
   AIChatTextareaProps,
   AIChatTextareaSubmit,
-  AIGlobalCommandProps,
-  AIGlobalCommandRefProps,
   AIInputInnerFeatureEnum,
-  AIManualAdditionProps,
   FileToChatQuestionList,
   FooterLeftTypesComponentProps,
   QSInputTextareaProps,
@@ -23,7 +20,7 @@ import {
 import { Input } from 'antd'
 import { YakitButton } from '@/components/yakitUI/YakitButton/YakitButton'
 import { OutlineArrowupIcon, OutlineAtsymbolIcon, OutlineCodeIcon, OutlineHandIcon } from '@/assets/icon/outline'
-import { useCreation, useInViewport, useMemoizedFn, useUpdateEffect } from 'ahooks'
+import { useCreation, useInViewport, useMemoizedFn } from 'ahooks'
 import { TextAreaRef } from 'antd/lib/input/TextArea'
 import classNames from 'classnames'
 import styles from './template.module.scss'
@@ -49,17 +46,12 @@ import { isString } from 'lodash'
 import OpenFileDropdown, { OpenFileDropdownItem } from '../aiChatWelcome/OpenFileDropdown/OpenFileDropdown'
 import { UploadFileButton } from '@/pages/ai-re-act/aiReActChat/AIReActComponent'
 import { insertAtCurrentPosition } from '../components/aiMilkdownInput/customPlugin'
-import { YakitPopover } from '@/components/yakitUI/YakitPopover/YakitPopover'
-import { YakitInput } from '@/components/yakitUI/YakitInput/YakitInput'
 import useAIAgentDispatcher from '../useContext/useDispatcher'
 import { YakitTag } from '@/components/yakitUI/YakitTag/YakitTag'
 import useChatIPCStore from '../useContext/ChatIPCContent/useStore'
-import useChatIPCDispatcher from '../useContext/ChatIPCContent/useDispatcher'
-import { AIInputEventSyncTypeEnum } from '@/pages/ai-re-act/hooks/grpcApi'
-import { randomString } from '@/utils/randomUtil'
-import { YakitPopconfirm } from '@/components/yakitUI/YakitPopconfirm/YakitPopconfirm'
 import useAIGlobalConfig from '@/pages/ai-re-act/hooks/useAIGlobalConfig'
 import { YakitSpin } from '@/components/yakitUI/YakitSpin/YakitSpin'
+import { AIGlobalCommandPopover, AIManualAdditionPopover } from '@/pages/ai-re-act/aiReActTaskChat/AIReActTaskChat'
 
 /** @name AI-Agent专用Textarea组件,行高为20px */
 export const QSInputTextarea: React.FC<QSInputTextareaProps & RefAttributes<TextAreaRef>> = memo(
@@ -100,7 +92,6 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
     const { chatIPCData } = useChatIPCStore()
     const execute = useCreation(() => chatIPCData.execute, [chatIPCData.execute])
 
-    const [visible, setVisible] = useState<boolean>(false)
     const [manualAdditionVisible, setManualAdditionVisible] = useState<boolean>(false)
 
     const footerLeftTypes: FooterLeftTypesComponentProps[] = useCreation(() => {
@@ -315,25 +306,10 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
       editorMilkdown.current?.action(callCommand<string>(insertAtCurrentPosition.key, '@'))
     })
 
-    //#region AI全局指令相关逻辑
     const [aiGlobalConfigData, event] = useAIGlobalConfig()
 
     const aiGlobalConfig = useCreation(() => aiGlobalConfigData.aiGlobalConfig, [aiGlobalConfigData.aiGlobalConfig])
     const updateLoading = useCreation(() => aiGlobalConfigData.updateLoading, [aiGlobalConfigData.updateLoading])
-
-    const onSave = useMemoizedFn((prompt: string) => {
-      setVisible(false)
-      event.setAIGlobalConfig({ AIPresetPrompt: prompt })
-    })
-    const aiGlobalCommandRef = useRef<AIGlobalCommandRefProps>({ value: '' })
-    const onGlobalCommandVisibleChange = useMemoizedFn((visible: boolean) => {
-      if (!visible) {
-        onSave(aiGlobalCommandRef.current?.value || '')
-      } else {
-        setVisible(true)
-      }
-    })
-    //#endregion
     return (
       <div
         className={classNames(
@@ -348,28 +324,14 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
         ref={dropRef}
       >
         {isHovering && <div className={styles['drag-hint']}>松开以添加到对话</div>}
-        <YakitPopover
-          visible={visible}
-          content={<AIGlobalCommand ref={aiGlobalCommandRef} onCancel={() => setVisible(false)} onSave={onSave} />}
-          destroyTooltipOnHide={true}
-          onVisibleChange={onGlobalCommandVisibleChange}
-          trigger={'click'}
-        >
-          <div
-            onClick={(e) => {
-              e.stopPropagation()
-              setVisible(true)
-            }}
-            className={styles['code-btn-wrapper']}
-          >
-            <YakitSpin spinning={updateLoading} size="small">
-              <YakitTag color="purple" size="small" border={false} fullRadius className={styles['preset-prompt-tag']}>
-                <OutlineCodeIcon className={styles['code-icon']} />
-                <span className="content-ellipsis">{aiGlobalConfig.AIPresetPrompt || '自定义AI全局指令'}</span>
-              </YakitTag>
-            </YakitSpin>
-          </div>
-        </YakitPopover>
+        <AIGlobalCommandPopover childrenClass={styles['code-btn-wrapper']}>
+          <YakitSpin spinning={updateLoading} size="small">
+            <YakitTag color="purple" size="small" border={false} fullRadius className={styles['preset-prompt-tag']}>
+              <OutlineCodeIcon className={styles['code-icon']} />
+              <span className="content-ellipsis">{aiGlobalConfig.AIPresetPrompt || '自定义AI全局指令'}</span>
+            </YakitTag>
+          </YakitSpin>
+        </AIGlobalCommandPopover>
         <div className={classNames(styles['textarea-wrapper'])} onKeyDown={handleTextareaKeyDown}>
           <AIMilkdownInput
             defaultValue={defaultValue}
@@ -392,12 +354,7 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
                   <UploadFileButton title="打开文件夹" className={styles['btn-base']} />
                 </OpenFileDropdown>
 
-                <YakitPopover
-                  visible={manualAdditionVisible}
-                  content={<AIManualAddition onCancel={() => setManualAdditionVisible(false)} />}
-                  onVisibleChange={setManualAdditionVisible}
-                  trigger={'click'}
-                >
+                <AIManualAdditionPopover visible={manualAdditionVisible} setVisible={setManualAdditionVisible}>
                   <YakitButton
                     type="text2"
                     radius="50%"
@@ -409,7 +366,7 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
                       e.stopPropagation()
                     }}
                   />
-                </YakitPopover>
+                </AIManualAdditionPopover>
               </div>
             )}
 
@@ -437,143 +394,3 @@ export const AIChatTextarea: React.FC<AIChatTextareaProps> = memo(
     )
   }),
 )
-
-const AIGlobalCommand: React.FC<AIGlobalCommandProps> = React.memo(
-  forwardRef((props, ref) => {
-    const { onCancel, onSave } = props
-    const [aiGlobalConfigData] = useAIGlobalConfig()
-    const [prompt, setPrompt] = useState<string>(aiGlobalConfigData?.aiGlobalConfig.AIPresetPrompt || '')
-    useImperativeHandle(ref, () => ({ value: prompt }), [prompt])
-
-    return (
-      <div className={styles['ai-global-command']} onClick={(e) => e.stopPropagation()}>
-        <div className={styles['ai-global-command-heard']}>全局命令</div>
-        <YakitInput.TextArea
-          rows={5}
-          value={prompt}
-          onChange={(e) => setPrompt(e.target.value)}
-          isShowResize={false}
-          onPressEnter={() => onSave(prompt)}
-          placeholder="全局命令将对所有会话生效..."
-          maxLength={500}
-          showCount
-        />
-        <div className={styles['ai-global-command-footer']}>
-          <YakitButton type="outline2" onClick={onCancel}>
-            取消
-          </YakitButton>
-          <YakitButton
-            onClick={() => {
-              onSave(prompt)
-            }}
-          >
-            保存
-          </YakitButton>
-        </div>
-      </div>
-    )
-  }),
-)
-
-const AIManualAddition: React.FC<AIManualAdditionProps> = React.memo((props) => {
-  const { onCancel } = props
-  const { handleSendSyncMessage, chatIPCEvents } = useChatIPCDispatcher()
-  const { chatIPCData, syncIdInfoMap } = useChatIPCStore()
-  const [prompt, setPrompt] = useState<string>()
-
-  const currentCoordinatorIdRef = useRef<string>('')
-  const syncIdOfAddToContext = useRef<string>('')
-  const syncIdOfAddAndReExecute = useRef<string>('')
-
-  const taskStatus = useCreation(() => chatIPCData?.taskStatus, [chatIPCData?.taskStatus])
-
-  useUpdateEffect(() => {
-    if (!taskStatus.loading && currentCoordinatorIdRef.current) {
-      onSendRecover(currentCoordinatorIdRef.current)
-    }
-  }, [taskStatus.loading])
-
-  useEffect(() => {
-    if (
-      (syncIdOfAddToContext.current && !syncIdInfoMap?.get(syncIdOfAddToContext.current)) ||
-      (syncIdOfAddAndReExecute.current && !syncIdInfoMap?.get(syncIdOfAddAndReExecute.current))
-    ) {
-      onCancel()
-      setPrompt('')
-      if (syncIdOfAddToContext.current) syncIdOfAddToContext.current = ''
-      if (syncIdOfAddAndReExecute.current) syncIdOfAddAndReExecute.current = ''
-    }
-  }, [syncIdInfoMap])
-
-  const onAddAndReExecute = useMemoizedFn(() => {
-    if (!prompt?.trim()) return
-    // 加入上下文后，停止任务再恢复任务
-    syncIdOfAddAndReExecute.current = randomString(8)
-    onAddToContext(syncIdOfAddAndReExecute.current)
-    const info = chatIPCEvents.fetchTaskChatID()
-    const taskId = info?.taskID
-    const coordinatorId = info?.coordinatorId
-    if (!coordinatorId) return
-    currentCoordinatorIdRef.current = coordinatorId
-    chatIPCEvents.handleCancelLoadingChange('task', true)
-    if (taskStatus?.loading && taskId) {
-      // 选停止当前任务，等待任务停止成功后，再发送恢复的数据
-      handleSendSyncMessage({
-        syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_REACT_CANCEL_TASK,
-        SyncJsonInput: JSON.stringify({ task_id: taskId }),
-      })
-    } else {
-      onSendRecover(coordinatorId)
-    }
-  })
-  const onSendRecover = useMemoizedFn((coordinatorId: string) => {
-    handleSendSyncMessage({
-      syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_RECOVERY_PLAN_AND_EXEC,
-      SyncJsonInput: JSON.stringify({ coordinator_id: coordinatorId }),
-    })
-    currentCoordinatorIdRef.current = ''
-  })
-  const onAddToContext = useMemoizedFn((syncID: string) => {
-    if (!prompt?.trim()) return
-    handleSendSyncMessage({
-      syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_USER_INTERVENTION,
-      SyncJsonInput: JSON.stringify({ content: prompt }),
-      syncID: syncID,
-    })
-  })
-  return (
-    <div className={styles['ai-manual-addition']} onClick={(e) => e.stopPropagation()}>
-      <div className={styles['ai-manual-addition-heard']}>手动介入</div>
-      <YakitInput.TextArea
-        rows={5}
-        value={prompt}
-        onChange={(e) => setPrompt(e.target.value)}
-        isShowResize={false}
-        placeholder="输入补充内容会加入上下文影响任务执行"
-      />
-      <div className={styles['ai-manual-addition-footer']}>
-        <YakitPopconfirm title="如果当前有任务正在执行,确认后会停止当前任务并重新执行" onConfirm={onAddAndReExecute}>
-          <YakitButton
-            type="outline2"
-            onClick={(e) => e.stopPropagation()}
-            loading={!!syncIdInfoMap?.get(syncIdOfAddAndReExecute.current)}
-            className={styles['add-and-reexecute-btn']}
-            disabled={!!syncIdInfoMap?.get(syncIdOfAddToContext.current)}
-          >
-            加入并重新执行
-          </YakitButton>
-        </YakitPopconfirm>
-        <YakitButton
-          onClick={() => {
-            syncIdOfAddToContext.current = randomString(8)
-            onAddToContext(syncIdOfAddToContext.current)
-          }}
-          loading={!!syncIdInfoMap?.get(syncIdOfAddToContext.current)}
-          disabled={!!syncIdInfoMap?.get(syncIdOfAddAndReExecute.current)}
-        >
-          加入上下文
-        </YakitButton>
-      </div>
-    </div>
-  )
-})
