+++ app/renderer/src/main/src/pages/ai-re-act/aiReActTaskChat/AIReActTaskChat.tsx
@@ -1,5 +1,10 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React, { forwardRef, useEffect, useImperativeHandle, useRef, useState } from 'react'
 import {
+  AIGlobalCommandPopoverProps,
+  AIGlobalCommandProps,
+  AIGlobalCommandRefProps,
+  AIManualAdditionPopoverProps,
+  AIManualAdditionProps,
   AIReActTaskChatContentProps,
   AIReActTaskChatLeftSideProps,
   AIReActTaskChatProps,
@@ -17,7 +22,9 @@ import { AIReActTaskChatReview } from '@/pages/ai-agent/aiAgentChat/AIAgentChat'
 import {
   OutlineArrowscollapseIcon,
   OutlineArrowsexpandIcon,
+  OutlineCodeIcon,
   OutlineExitIcon,
+  OutlineHandIcon,
   OutlinePlay2Icon,
   OutlinePlussmIcon,
   OutlinePositionIcon,
@@ -39,6 +46,9 @@ import { YakitRoute } from '@/enums/yakitRoute'
 import { AIMentionCommandParams } from '@/pages/ai-agent/components/aiMilkdownInput/aiMilkdownMention/aiMentionPlugin'
 import { YakitResizeBox } from '@/components/yakitUI/YakitResizeBox/YakitResizeBox'
 import { useI18nNamespaces } from '@/i18n/useI18nNamespaces'
+import { YakitPopover } from '@/components/yakitUI/YakitPopover/YakitPopover'
+import useAIGlobalConfig from '../hooks/useAIGlobalConfig'
+import { YakitInput } from '@/components/yakitUI/YakitInput/YakitInput'
 
 const AIReActTaskChat: React.FC<AIReActTaskChatProps> = React.memo((props) => {
   const { setShowFreeChat, setTimeLine } = props
@@ -200,9 +210,11 @@ const AIReActTaskRecommend: React.FC<AIReActTaskRecommendProps> = React.memo((pr
 
 const AIReActTaskChatContent: React.FC<AIReActTaskChatContentProps> = React.memo((props) => {
   const { reviewInfo, planReviewTreeKeywordsMap, chatIPCData } = useChatIPCStore()
-
+  const { t } = useI18nNamespaces(['aiAgent'])
   const { activeChat } = useAIAgentStore()
   const { taskChat } = chatIPCData
+
+  const [visible, setVisible] = useState<boolean>(false)
 
   const { handleSendSyncMessage, chatIPCEvents } = useChatIPCDispatcher()
 
@@ -322,6 +334,26 @@ const AIReActTaskChatContent: React.FC<AIReActTaskChatContentProps> = React.memo
       ) : (
         streams.length > 0 && (
           <div className={styles['footer']}>
+            {chatIPCData.execute && (
+              <AIManualAdditionPopover>
+                <YakitButton
+                  type="outline2"
+                  radius="28px"
+                  icon={<OutlineHandIcon />}
+                  onClick={(e) => {
+                    e.stopPropagation()
+                  }}
+                >
+                  {t('AIReActTaskChatContent.humanIntervention')}
+                </YakitButton>
+              </AIManualAdditionPopover>
+            )}
+
+            <AIGlobalCommandPopover>
+              <YakitButton icon={<OutlineCodeIcon />} radius="28px" type="outline2">
+                {t('AIReActTaskChatContent.globalDirective')}
+              </YakitButton>
+            </AIGlobalCommandPopover>
             {!!getTaskId() && (
               <>
                 <AIRenderTaskFooterExtra onExtraAction={onExtraAction} />
@@ -341,7 +373,206 @@ const AIReActTaskChatContent: React.FC<AIReActTaskChatContentProps> = React.memo
     </>
   )
 })
+export const AIManualAdditionPopover: React.FC<AIManualAdditionPopoverProps> = React.memo((props) => {
+  const { children } = props
+  const [manualAdditionVisible, setManualAdditionVisible] = useControllableValue<boolean>(props, {
+    defaultValue: false,
+    valuePropName: 'visible',
+    trigger: 'setVisible',
+  })
 
+  return (
+    <YakitPopover
+      visible={manualAdditionVisible}
+      content={<AIManualAddition onCancel={() => setManualAdditionVisible(false)} />}
+      onVisibleChange={setManualAdditionVisible}
+      trigger={'click'}
+    >
+      {children}
+    </YakitPopover>
+  )
+})
+
+const AIManualAddition: React.FC<AIManualAdditionProps> = React.memo((props) => {
+  const { onCancel } = props
+  const { handleSendSyncMessage, chatIPCEvents } = useChatIPCDispatcher()
+  const { chatIPCData, syncIdInfoMap } = useChatIPCStore()
+  const [prompt, setPrompt] = useState<string>()
+
+  const currentCoordinatorIdRef = useRef<string>('')
+  const syncIdOfAddToContext = useRef<string>('')
+  const syncIdOfAddAndReExecute = useRef<string>('')
+
+  const taskStatus = useCreation(() => chatIPCData?.taskStatus, [chatIPCData?.taskStatus])
+
+  useUpdateEffect(() => {
+    if (!taskStatus.loading && currentCoordinatorIdRef.current) {
+      onSendRecover(currentCoordinatorIdRef.current)
+    }
+  }, [taskStatus.loading])
+
+  useEffect(() => {
+    if (
+      (syncIdOfAddToContext.current && !syncIdInfoMap?.get(syncIdOfAddToContext.current)) ||
+      (syncIdOfAddAndReExecute.current && !syncIdInfoMap?.get(syncIdOfAddAndReExecute.current))
+    ) {
+      onCancel()
+      setPrompt('')
+      if (syncIdOfAddToContext.current) syncIdOfAddToContext.current = ''
+      if (syncIdOfAddAndReExecute.current) syncIdOfAddAndReExecute.current = ''
+    }
+  }, [syncIdInfoMap])
+
+  const onAddAndReExecute = useMemoizedFn(() => {
+    if (!prompt?.trim()) return
+    // 加入上下文后，停止任务再恢复任务
+    syncIdOfAddAndReExecute.current = randomString(8)
+    onAddToContext(syncIdOfAddAndReExecute.current)
+    const info = chatIPCEvents.fetchTaskChatID()
+    const taskId = info?.taskID
+    const coordinatorId = info?.coordinatorId
+    if (!coordinatorId) return
+    currentCoordinatorIdRef.current = coordinatorId
+    chatIPCEvents.handleCancelLoadingChange('task', true)
+    if (taskStatus?.loading && taskId) {
+      // 选停止当前任务，等待任务停止成功后，再发送恢复的数据
+      handleSendSyncMessage({
+        syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_REACT_CANCEL_TASK,
+        SyncJsonInput: JSON.stringify({ task_id: taskId }),
+      })
+    } else {
+      onSendRecover(coordinatorId)
+    }
+  })
+  const onSendRecover = useMemoizedFn((coordinatorId: string) => {
+    handleSendSyncMessage({
+      syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_RECOVERY_PLAN_AND_EXEC,
+      SyncJsonInput: JSON.stringify({ coordinator_id: coordinatorId }),
+    })
+    currentCoordinatorIdRef.current = ''
+  })
+  const onAddToContext = useMemoizedFn((syncID: string) => {
+    if (!prompt?.trim()) return
+    handleSendSyncMessage({
+      syncType: AIInputEventSyncTypeEnum.SYNC_TYPE_USER_INTERVENTION,
+      SyncJsonInput: JSON.stringify({ content: prompt }),
+      syncID: syncID,
+    })
+  })
+  return (
+    <div className={styles['ai-manual-addition']} onClick={(e) => e.stopPropagation()}>
+      <div className={styles['ai-manual-addition-heard']}>手动介入</div>
+      <YakitInput.TextArea
+        rows={5}
+        value={prompt}
+        onChange={(e) => setPrompt(e.target.value)}
+        isShowResize={false}
+        placeholder="输入补充内容会加入上下文影响任务执行"
+        maxLength={500}
+        showCount
+      />
+      <div className={styles['ai-manual-addition-footer']}>
+        <YakitPopconfirm title="如果当前有任务正在执行,确认后会停止当前任务并重新执行" onConfirm={onAddAndReExecute}>
+          <YakitButton
+            type="outline2"
+            onClick={(e) => e.stopPropagation()}
+            loading={!!syncIdInfoMap?.get(syncIdOfAddAndReExecute.current)}
+            className={styles['add-and-reexecute-btn']}
+            disabled={!!syncIdInfoMap?.get(syncIdOfAddToContext.current)}
+          >
+            加入并重新执行
+          </YakitButton>
+        </YakitPopconfirm>
+        <YakitButton
+          onClick={() => {
+            syncIdOfAddToContext.current = randomString(8)
+            onAddToContext(syncIdOfAddToContext.current)
+          }}
+          loading={!!syncIdInfoMap?.get(syncIdOfAddToContext.current)}
+          disabled={!!syncIdInfoMap?.get(syncIdOfAddAndReExecute.current)}
+        >
+          加入上下文
+        </YakitButton>
+      </div>
+    </div>
+  )
+})
+export const AIGlobalCommandPopover: React.FC<AIGlobalCommandPopoverProps> = React.memo((props) => {
+  const { children, childrenClass } = props
+  const [visible, setVisible] = useState<boolean>(false)
+  //#region AI全局指令相关逻辑
+  const [_, event] = useAIGlobalConfig()
+
+  const onSave = useMemoizedFn((prompt: string) => {
+    setVisible(false)
+    event.setAIGlobalConfig({ AIPresetPrompt: prompt })
+  })
+  const aiGlobalCommandRef = useRef<AIGlobalCommandRefProps>({ value: '' })
+  const onGlobalCommandVisibleChange = useMemoizedFn((visible: boolean) => {
+    if (!visible) {
+      onSave(aiGlobalCommandRef.current?.value || '')
+    } else {
+      setVisible(true)
+    }
+  })
+  //#endregion
+  return (
+    <YakitPopover
+      visible={visible}
+      content={<AIGlobalCommand ref={aiGlobalCommandRef} onCancel={() => setVisible(false)} onSave={onSave} />}
+      destroyTooltipOnHide={true}
+      onVisibleChange={onGlobalCommandVisibleChange}
+      trigger={'click'}
+    >
+      <div
+        onClick={(e) => {
+          e.stopPropagation()
+          setVisible(true)
+        }}
+        className={classNames(childrenClass)}
+      >
+        {children}
+      </div>
+    </YakitPopover>
+  )
+})
+const AIGlobalCommand: React.FC<AIGlobalCommandProps> = React.memo(
+  forwardRef((props, ref) => {
+    const { t } = useI18nNamespaces(['aiAgent', 'yakitUi'])
+    const { onCancel, onSave } = props
+    const [aiGlobalConfigData] = useAIGlobalConfig()
+    const [prompt, setPrompt] = useState<string>(aiGlobalConfigData?.aiGlobalConfig.AIPresetPrompt || '')
+    useImperativeHandle(ref, () => ({ value: prompt }), [prompt])
+
+    return (
+      <div className={styles['ai-global-command']} onClick={(e) => e.stopPropagation()}>
+        <div className={styles['ai-global-command-heard']}>{t('AIReActTaskChatContent.humanIntervention')}</div>
+        <YakitInput.TextArea
+          rows={5}
+          value={prompt}
+          onChange={(e) => setPrompt(e.target.value)}
+          isShowResize={false}
+          onPressEnter={() => onSave(prompt)}
+          placeholder="全局命令将对所有会话生效..."
+          maxLength={500}
+          showCount
+        />
+        <div className={styles['ai-global-command-footer']}>
+          <YakitButton type="outline2" onClick={onCancel}>
+            {t('YakitButton.cancel')}
+          </YakitButton>
+          <YakitButton
+            onClick={() => {
+              onSave(prompt)
+            }}
+          >
+            {t('YakitButton.save')}
+          </YakitButton>
+        </div>
+      </div>
+    )
+  }),
+)
 const AIRenderTaskFooterExtra: React.FC<AIRenderTaskFooterExtraProps> = React.memo((props) => {
   const { onExtraAction, btnProps, subTaskBtnProps, children } = props
   const { t } = useI18nNamespaces(['aiAgent'])
